LayoutTests/http/tests/security/mixedContent/mainframe-onload-after-iframe-block.https-expected.txt

CONSOLE MESSAGE: [blocked] The page at https://127.0.0.1:8443/security/mixedContent/mainframe-onload-after-iframe-block.https.html requested insecure content from http://127.0.0.1:8443/frame-b.html. This content was blocked and must be served over HTTPS.

CONSOLE MESSAGE: Onload Triggered!
